Setup navigation and theme for wearos
