+++ components/wearable/setup/api/build.gradle.kts
@@ -1,0 +1,15 @@
+plugins {
+    id("flipper.lint")
+    id("flipper.android-compose")
+}
+
+dependencies {
+    implementation(projects.components.core.ui.navigation)
+
+    // Compose
+    implementation(libs.compose.ui)
+    implementation(libs.compose.tooling)
+    implementation(libs.compose.wear.foundation)
+    implementation(libs.compose.wear.material)
+    implementation(libs.compose.wear.navigation)
+}
