+++ components/wearable/setup/api/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.flipperdevices.wearable.setup.api" />
