+++ components/wearable/setup/api/src/main/java/com/flipperdevices/wearable/setup/api/SetupApi.kt
@@ -1,0 +1,11 @@
+package com.flipperdevices.wearable.setup.api
+
+import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
+
+interface SetupApi : AggregateFeatureEntry {
+    fun start(): String
+
+    companion object {
+        const val ROUTE = "@wearossetup"
+    }
+}
