+++ components/wearable/setup/impl/build.gradle.kts
@@ -1,0 +1,25 @@
+plugins {
+    id("flipper.lint")
+    id("flipper.android-compose")
+    id("com.squareup.anvil")
+    id("kotlin-kapt")
+}
+
+dependencies {
+    implementation(projects.components.wearable.setup.api)
+
+    implementation(projects.components.core.di)
+    implementation(projects.components.core.ui.theme)
+    implementation(projects.components.core.ui.navigation)
+
+    // Compose
+    implementation(libs.compose.ui)
+    implementation(libs.compose.tooling)
+    implementation(libs.compose.wear.foundation)
+    implementation(libs.compose.wear.material)
+    implementation(libs.compose.wear.navigation)
+
+    // Dagger deps
+    implementation(libs.dagger)
+    kapt(libs.dagger.kapt)
+}
