+++ components/wearable/setup/impl/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.flipperdevices.wearable.setup.impl" />
