+++ components/wearable/theme/build.gradle.kts
@@ -1,0 +1,14 @@
+plugins {
+    id("flipper.lint")
+    id("flipper.android-compose")
+}
+
+dependencies {
+    implementation(projects.components.core.ui.theme)
+
+    // Compose
+    implementation(libs.compose.ui)
+    implementation(libs.compose.tooling)
+    implementation(libs.compose.wear.foundation)
+    implementation(libs.compose.wear.material)
+}
