+++ instances/wearable/src/main/res/values/wear.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources xmlns:tools="http://schemas.android.com/tools" tools:keep="@array/android_wear_capabilities">
+    <string-array name="android_wear_capabilities">
+        <item>verify_remote_flipper_wear_app</item>
+    </string-array>
+</resources>
